backend/requirements.txt
pandas>=2.0
fastapi
uvicorn
python-dotenv
openai
pydantic
pydantic-settings
firebase-admin
pymongo
httpx
python-multipart
chromadb
pymupdf

# TTS — Coqui XTTS-v2 (local inference)
TTS
torch
soundfile
